refactor: address all follow-up items from code review
Performance:
- Replace 3x index.query() with SQL aggregation (countByType, sporeCountsByObservationType)
- Add 4 new tests for aggregation methods (503 total)

Component decomposition:
- Split ConfigForm from 753 → 307 lines (6 section components extracted)
- Split Dashboard from 442 → 59 lines (stat cards + quick actions extracted)
- Extract shared PageLoading component, replace copy-pasted loading/error in 3 pages

Shared utilities:
- Extract formatUptime/formatTimeAgo to lib/format.ts
- Fix DryRunResult type to match backend curation response shape

UX improvements:
- Add visual progress bar to OperationButton
- Wire OpenVault in Layout sidebar to real vault URIs from stats API
