+++ ui/src/components/config/CaptureSection.tsx
@@ -1,0 +1,58 @@
+import type { MycoConfig } from '../../hooks/use-config';
+import { ConfigSection } from './ConfigSection';
+import { Input } from '../ui/input';
+import { Field, numChange } from './config-helpers';
+
+interface CaptureSectionProps {
+  capture: MycoConfig['capture'];
+  isDirty: boolean;
+  updateCapture: (key: string, value: number) => void;
+}
+
+export function CaptureSection({ capture, isDirty, updateCapture }: CaptureSectionProps) {
+  return (
+    <ConfigSection
+      title="Capture"
+      description="Event buffering and token budgets for LLM processing"
+      isDirty={isDirty}
+    >
+      <div className="grid gap-4 sm:grid-cols-2">
+        <Field label="Buffer Max Events">
+          <Input
+            type="number"
+            value={capture.buffer_max_events}
+            onChange={numChange(updateCapture, 'buffer_max_events')}
+          />
+        </Field>
+        <Field label="Extraction Max Tokens">
+          <Input
+            type="number"
+            value={capture.extraction_max_tokens}
+            onChange={numChange(updateCapture, 'extraction_max_tokens')}
+          />
+        </Field>
+        <Field label="Summary Max Tokens">
+          <Input
+            type="number"
+            value={capture.summary_max_tokens}
+            onChange={numChange(updateCapture, 'summary_max_tokens')}
+          />
+        </Field>
+        <Field label="Title Max Tokens">
+          <Input
+            type="number"
+            value={capture.title_max_tokens}
+            onChange={numChange(updateCapture, 'title_max_tokens')}
+          />
+        </Field>
+        <Field label="Classification Max Tokens">
+          <Input
+            type="number"
+            value={capture.classification_max_tokens}
+            onChange={numChange(updateCapture, 'classification_max_tokens')}
+          />
+        </Field>
+      </div>
+    </ConfigSection>
+  );
+}
