+++ ui/src/pages/Operations.tsx
@@ -4,7 +4,6 @@ import {
   Sparkles,
   Brain,
   Clock,
-  Database,
   Wrench,
 } from 'lucide-react';
 import { useDaemon } from '../hooks/use-daemon';
@@ -20,6 +19,7 @@ import {
   SelectValue,
 } from '../components/ui/select';
 import { Button } from '../components/ui/button';
+import { PageLoading } from '../components/ui/page-loading';
 
 /* ---------- Constants ---------- */
 
@@ -29,124 +29,109 @@ const DEFAULT_DIGEST_TIER = '3000';
 /* ---------- Operations Page ---------- */
 
 export default function Operations() {
-  const { data: stats, isLoading, isError } = useDaemon();
-
-  if (isLoading) {
-    return (
-      <div className="flex h-full items-center justify-center">
-        <div className="flex flex-col items-center gap-3 text-muted-foreground">
-          <RefreshCw className="h-6 w-6 animate-spin" />
-          <span className="text-sm">Connecting to daemon...</span>
-        </div>
-      </div>
-    );
-  }
-
-  if (isError || !stats) {
-    return (
-      <div className="flex h-full items-center justify-center">
-        <div className="flex flex-col items-center gap-3 text-muted-foreground">
-          <Database className="h-6 w-6 opacity-50" />
-          <span className="text-sm">Unable to reach daemon</span>
-          <span className="text-xs opacity-60">Check that the daemon is running</span>
-        </div>
-      </div>
-    );
-  }
+  const { data: stats, isLoading, isError, error } = useDaemon();
 
   return (
-    <div className="flex flex-col gap-6 p-6">
-      <div className="flex items-center gap-3">
-        <Wrench className="h-5 w-5 text-primary" />
-        <h1 className="text-xl font-semibold">Operations</h1>
-      </div>
+    <PageLoading
+      isLoading={isLoading}
+      error={isError ? (error instanceof Error ? error : new Error('Unable to reach daemon')) : null}
+      loadingText="Connecting to daemon..."
+    >
+      {stats && (
+        <div className="flex flex-col gap-6 p-6">
+          <div className="flex items-center gap-3">
+            <Wrench className="h-5 w-5 text-primary" />
+            <h1 className="text-xl font-semibold">Operations</h1>
+          </div>
 
-      <div className="grid gap-6 lg:grid-cols-2">
-        {/* Curation */}
-        <Card>
-          <CardHeader>
-            <CardTitle className="flex items-center gap-2 text-sm">
-              <Sparkles className="h-4 w-4 text-primary" />
-              Curation
-            </CardTitle>
-            <CardDescription>
-              Find and consolidate duplicate spores in the vault.
-              Preview shows which pairs would be merged; execute performs the consolidation.
-            </CardDescription>
-          </CardHeader>
-          <CardContent>
-            <CurationPanel />
-          </CardContent>
-        </Card>
+          <div className="grid gap-6 lg:grid-cols-2">
+            {/* Curation */}
+            <Card>
+              <CardHeader>
+                <CardTitle className="flex items-center gap-2 text-sm">
+                  <Sparkles className="h-4 w-4 text-primary" />
+                  Curation
+                </CardTitle>
+                <CardDescription>
+                  Find and consolidate duplicate spores in the vault.
+                  Preview shows which pairs would be merged; execute performs the consolidation.
+                </CardDescription>
+              </CardHeader>
+              <CardContent>
+                <CurationPanel />
+              </CardContent>
+            </Card>
 
-        {/* Rebuild Index */}
-        <Card>
-          <CardHeader>
-            <CardTitle className="flex items-center gap-2 text-sm">
-              <RefreshCw className="h-4 w-4 text-primary" />
-              Rebuild Index
-            </CardTitle>
-            <CardDescription>
-              Reindex all vault notes for full-text search and vector embeddings.
-            </CardDescription>
-          </CardHeader>
-          <CardContent>
-            <div className="space-y-4">
-              <OperationButton
-                label="Rebuild"
-                endpoint="/rebuild"
-                icon={<RefreshCw className="h-4 w-4" />}
-                description="Rebuilds FTS5 and vector indexes from all vault notes."
-              />
-              <div className="flex items-center gap-4 text-sm">
-                <div className="flex items-center gap-2">
-                  <span className="text-muted-foreground">FTS entries:</span>
-                  <Badge variant="secondary" className="font-mono text-xs">
-                    {stats.index.fts_entries}
-                  </Badge>
+            {/* Rebuild Index */}
+            <Card>
+              <CardHeader>
+                <CardTitle className="flex items-center gap-2 text-sm">
+                  <RefreshCw className="h-4 w-4 text-primary" />
+                  Rebuild Index
+                </CardTitle>
+                <CardDescription>
+                  Reindex all vault notes for full-text search and vector embeddings.
+                </CardDescription>
+              </CardHeader>
+              <CardContent>
+                <div className="space-y-4">
+                  <OperationButton
+                    label="Rebuild"
+                    endpoint="/rebuild"
+                    icon={<RefreshCw className="h-4 w-4" />}
+                    description="Rebuilds FTS5 and vector indexes from all vault notes."
+                  />
+                  <div className="flex items-center gap-4 text-sm">
+                    <div className="flex items-center gap-2">
+                      <span className="text-muted-foreground">FTS entries:</span>
+                      <Badge variant="secondary" className="font-mono text-xs">
+                        {stats.index.fts_entries}
+                      </Badge>
+                    </div>
+                    <div className="flex items-center gap-2">
+                      <span className="text-muted-foreground">Vectors:</span>
+                      <Badge variant="secondary" className="font-mono text-xs">
+                        {stats.index.vector_count}
+                      </Badge>
+                    </div>
+                  </div>
                 </div>
-                <div className="flex items-center gap-2">
-                  <span className="text-muted-foreground">Vectors:</span>
-                  <Badge variant="secondary" className="font-mono text-xs">
-                    {stats.index.vector_count}
+              </CardContent>
+            </Card>
+
+            {/* Manual Digest */}
+            <DigestSection
+              digestEnabled={!!stats.digest}
+              metabolismState={stats.digest?.metabolism_state ?? null}
+              substrateQueue={stats.digest?.substrate_queue ?? 0}
+            />
+
+            {/* Reprocess — Coming Soon */}
+            <Card>
+              <CardHeader>
+                <CardTitle className="flex items-center gap-2 text-sm">
+                  <Clock className="h-4 w-4 text-primary" />
+                  Reprocess
+                  <Badge variant="outline" className="ml-1 text-xs">
+                    Coming Soon
                   </Badge>
-                </div>
-              </div>
-            </div>
-          </CardContent>
-        </Card>
-
-        {/* Manual Digest */}
-        <DigestSection
-          digestEnabled={!!stats.digest}
-          metabolismState={stats.digest?.metabolism_state ?? null}
-          substrateQueue={stats.digest?.substrate_queue ?? 0}
-        />
-
-        {/* Reprocess — Coming Soon */}
-        <Card>
-          <CardHeader>
-            <CardTitle className="flex items-center gap-2 text-sm">
-              <Clock className="h-4 w-4 text-primary" />
-              Reprocess
-              <Badge variant="outline" className="ml-1 text-xs">
-                Coming Soon
-              </Badge>
-            </CardTitle>
-            <CardDescription>
-              Re-extract observations and regenerate summaries for past sessions.
-              This feature is being developed.
-            </CardDescription>
-          </CardHeader>
-          <CardContent>
-            <Button variant="outline" size="sm" disabled className="gap-2 opacity-50">
-              <Clock className="h-4 w-4" />
-              Reprocess Sessions
-            </Button>
-          </CardContent>
-        </Card>
-      </div>
-    </div>
+                </CardTitle>
+                <CardDescription>
+                  Re-extract observations and regenerate summaries for past sessions.
+                  This feature is being developed.
+                </CardDescription>
+              </CardHeader>
+              <CardContent>
+                <Button variant="outline" size="sm" disabled className="gap-2 opacity-50">
+                  <Clock className="h-4 w-4" />
+                  Reprocess Sessions
+                </Button>
+              </CardContent>
+            </Card>
+          </div>
+        </div>
+      )}
+    </PageLoading>
   );
 }
